requirements.txt
uvicorn~=0.17.6
zenlog~=1.1
pydantic~=1.9.1
fastapi~=0.79.0
pymongo~=4.1.1
python-dotenv~=0.20.0